refactor(plugins): relocate in ContractEditor folder
