+++ src/ContractEditor/plugins/README.md
@@ -85,69 +85,6 @@ return (
 );
 ```
 
-## ComputedPlugin
-
-### What it does?
-
-A custom Slate plugin for using non-editable, highlighted computed fields within a clause.
-Use ComputedPlugin to add a computed field and define schema for the addition of a computed field in the format defined by [slate.js](https://docs.slatejs.org/) as its used in the markdown-editor.
-
-### Usage
-
-```shell
-npm install @accordproject/cicero-ui
-```
-
-```js
-import { ComputedPlugin } from '@accordproject/cicero-ui';
-import { Editor } from 'slate-react'
-
-const plugins = [ComputedPlugin()];
-
-<Editor
-  ...
-  plugins={plugins}
-/>
-
-```
-
-### What it returns?
-
-It returns an object like this:
-
-```js
-{
-    name,
-    augmentSchema,
-    isEditable,
-    renderInline,
-}
-
-```
-
-where `augmentSchema`, `isEditable` and `renderInline` are functions.
-
-### augmentSchema function
-
-`augmentSchema` function returns a new schema for the addition of computed in the markdown editor in the format suggested by [slate.js](https://docs.slatejs.org/).
-
-### isEditable function
-
-`isEditable` function returns a boolean on checking whether the computed field in the markdown editor is editable or not.
-
-### renderInline function
-
-`renderInline` function renders a computed inline to the text in the editor by returning a span tag with the props.
-
-```js
-return (
-	<span id={id} {...attributes} className='computed'>
-		{" "}
-		{children}{" "}
-	</span>
-);
-```
-
 ## ClausePlugin
 
 A custom Slate plugin for embedding a clause node within a document. View the `demo` directory for more of an example.
@@ -235,61 +172,3 @@ where `augmentSchema`, `renderBlock`, `onChange`, `onPaste`, `findClauseNode`, `
 ### isClauseSupported
 
 `isClauseSupported` function returns a boolean which tells if the UI is valid.
-
-## ConditionalPlugin
-
-### What it does?
-
-A plugin to support variables corresponding to conditional blocks in templates enabling text which depends on a value of a Boolean variable in the model.
-
-### Usage
-
-```shell
-npm install @accordproject/cicero-ui
-```
-
-```js
-import { ConditionalPlugin } from '@accordproject/cicero-ui';
-import { Editor } from 'slate-react'
-
-const plugins = [ConditionalPlugin()];
-
-<Editor
-  ...
-  plugins={plugins}
-/>
-```
-
-### What it returns
-
-```js
-{
-  name,
-  augmentSchema,
-  isEditable,
-  renderInline,
-}
-```
-
-where `augmentSchema`, `isEditable` and `renderInline` are functions.
-
-### augmentSchema function
-
-`augmentSchema` function returns a new schema for the addition of conditional type in the markdown editor in the format suggested by [slate.js](https://docs.slatejs.org/).
-
-### isEditable function
-
-`isEditable` function returns a boolean on checking whether the conditional field in the markdown editor is editable or not.
-
-### renderInline function
-
-`renderInline` function renders a conditional inline to the text in the editor by returning a span tag with the props.
-
-```js
-return (
-	<span id={id} {...attributes} className='computed'>
-		{" "}
-		{children}{" "}
-	</span>
-);
-```
